[Nidhuna]refactor: type of field price changed from decimal to float field

+++ products/models.py
@@ -7,5 +7,5 @@ class Product(models.Model):
     title = models.CharField(max_length=100)
     image = models.CharField(max_length=500)
     quantity = models.IntegerField()
-    price = models.DecimalField(max_digits=20, decimal_places=3)
+    price = models.FloatField()
     description = models.CharField(max_length=10000)
